Define the numpy dtype-token aliases the field list was missing.
The wala/ML#865 remainder: `data/numpy.xml` carried only `float32
float64 int32 int64 uint8 bool bool_ string`, so every other dtype
spelling resolved to nothing. Paired with the earlier #865 half (an
API default is emitted only for a determinately absent argument), a
supplied-but-unmodeled spelling no longer coincidentally read the
float64 default; it read UNKNOWN, losing the dtype outright. This
defines the float, int, and string families so they resolve by name:
`np.float`/`float_`/`double` (float64), `np.single` (float32),
`np.int`/`long`/`int_`/`longlong`/`intp` (int64), `np.intc` (int32),
`np.str`/`str_`/`unicode`/`unicode_` (string).

Each spelling gets its OWN allocation and its OWN
`FIELD_REFERENCE_TO_DTYPE` entry, matching the existing per-name
pattern of `bool` and `bool_`. A first pass shared one allocation per
canonical type and relied on the resolver's cross-field identity match;
that match is subject to the wala/ML#753 collection-order
nondeterminism and resolved-or-degraded across JVMs, so the same
spelling read int64 in one fork and UNKNOWN in another. Per-name fields
resolve each token in its own field, deterministically (verified over
repeated fresh-JVM and warm-JVM runs).

The decisive witness is `np.int`: its truth is int64, which differs
from the old float64 default, so int64 proves resolution happened
rather than the default being restored. Two dependent fixtures move
onto still-unrepresentable spellings now that `np.int` resolves:
`TestDTypeAbsentVersusUnresolved` and the `np.eye` adopter arm switch
to `np.int16` (a current scalar type with no `DType` yet).

Scope, held to the reporter's families plus the string spellings found
alongside: `complex64`/`complex128` and `object` are deferred. Their
resolution entangles the wala/ML#816-blocked complex-division test and
did not stabilize, so they want their own investigation rather than
riding this change. Spellings whose dtypes have no `DType` yet (`int16`,
`int8`, `uint16`, `uint`, `float16`, `short`, `byte`) stay absent
pending an enum extension. The builtin `bool` sub-case the report also
lists did not reproduce: it is pinned positively in the witness, and a
construct that does fail it needs its own reduction.

+++ com.ibm.wala.cast.python.test/data/tf2_test_dtype_absent_vs_unresolved.py
@@ -20,14 +20,15 @@ def supplied_and_resolvable():
 
 
 def supplied_but_unresolved():
-    # Supplied, but spelled with an attribute the model file carries no field for, so nothing
-    # resolves. There is no evidence, and float64 here would be the defect: it is indistinguishable
+    # Supplied, but spelled with a dtype the model file carries no field for, so nothing resolves.
+    # There is no evidence, and float64 here would be the defect: it is indistinguishable
     # downstream from a dtype that genuinely resolved to float64.
     #
-    # `np.int` is the spelling the defect was reported with, kept deliberately: it is deprecated
-    # since NumPy 1.20 and removed in 1.24, so this fixture runs under NumPy < 1.24 (with a
-    # DeprecationWarning), matching the vendored subjects' era.
-    return np.ones([2, 3], dtype=np.int)
+    # `np.int16` is used because it is a current scalar type with no `DType` yet, so it stays
+    # unrepresentable BY DESIGN. The originally-reported `np.int` became RESOLVABLE when the
+    # wala/ML#865 field list grew; this arm needs a spelling that is genuinely unreadable, and
+    # int16's absence is the enforced one until the `DType` enum is extended.
+    return np.ones([2, 3], dtype=np.int16)
 
 
 def supplied_positionally():
@@ -60,7 +61,7 @@ def supplied_through_star(args):
 
 assert absent().dtype == np.float64
 assert supplied_and_resolvable().dtype == np.int32
-assert supplied_but_unresolved().dtype == np.int_
+assert supplied_but_unresolved().dtype == np.int16
 assert supplied_none().dtype == np.float64
 assert supplied_through_star(([2, 3],)).dtype == np.float64
 assert supplied_positionally().dtype == np.int32
